% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagMultiple.R
\name{lagMultiple}
\alias{lagMultiple}
\title{An extension on `diff` from base R}
\usage{
lagMultiple(x, k = 1)
}
\arguments{
\item{x}{A vector, single column matrix, or univariate time series}

\item{k}{(Optional) An integer vector containing a number of lags. Defaults
to 1}
}
\value{
Returns a data.frame of the lagged variable. The number of rows is
  the same as the length of the input vector. The number of columns is the
  number of lags to be used. Each column retains the name of the original
  variable and includes the number of lags used for that column.
}
\description{
An extension on `diff` from base R
}
\examples{
# Creating dummy data
x <- rnorm(100)
# for lags 1-5
lagMultiple(x, 1:5)
}
