% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_all_rows.R
\name{summarise_all_rows}
\alias{summarise_all_rows}
\title{A riff on `dplyr::summarise` that returns columns with values for all rows}
\usage{
summarise_all_rows(.data, ..., .groups = NULL)
}
\arguments{
\item{.data, ..., .groups}{Takes all the same arguments as
\code{\link[dplyr]{summarise}}}
}
\value{
A data.frame identical to `.data` with new columns representing the
  summarisations made after grouping. Values in the new columns are repeated
  as necessary to maintain the number of rows of the original data.frame.
}
\description{
A riff on `dplyr::summarise` that returns columns with values for all rows
}
\examples{
# Get mean mpg by number of cyl from mtcars
mtcars |>
  dplyr::group_by(cyl) |>
  summarise_all_rows(mean_mpg = mean(mpg))
}
