% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFill.R
\name{diffFill}
\alias{diffFill}
\title{An extension on `diff` from base R that fills blank space with NA}
\usage{
diffFill(x, lag = 1, differences = 1, name = NULL, ...)
}
\arguments{
\item{x}{a numeric vector OR matrix containing the values to be differenced}

\item{lag}{(Optional) an integer indicating which lag to use. If `x` is a
vector, `lag` can also be a vector. Defaults to 1}

\item{differences}{(Optional) an integer indicating the order of the
difference. If `x` is a vector, `differences` can also be a vector.
Defaults to 1}

\item{name}{(Optional) A name to be used in the lagged data.frame. Defaults
to the name of the variable passed to `x`. If that is not possible, `name`
will default to "X".}

\item{...}{(Optional) further arguments to be passed to or from methods}
}
\value{
Returns a vector of the same length as the input, otherwise behaves
  the same as \code{\link[base]{diff}}. If the input is a matrix, it returns
  a matrix.
}
\description{
An extension on `diff` from base R that fills blank space with NA
}
\examples{
# Creating dummy data
# for 1 lag and 2 differences
diffFill(rnorm(10), 1, 2)
}
