% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFill.R
\name{diffFill}
\alias{diffFill}
\title{An extension on `diff` from base R that fills blank space with NA}
\usage{
diffFill(x, lag = 1, differences = 1, ...)
}
\arguments{
\item{x}{a numeric vector or matrix containing the values to be differenced}

\item{lag}{(Optional) an integer indicating which lag to use. Defaults to 1}

\item{differences}{(Optional) an integer indicating the order of the
difference. Defaults to 1}

\item{...}{(Optional) further arguments to be passed to or from methods}
}
\value{
Returns a vector of the same length as the input, otherwise behaves
  the same as \code{\link[base]{diff}}
}
\description{
An extension on `diff` from base R that fills blank space with NA
}
\examples{
# Creating dummy data
x <- rnorm(10)
# for 1 lag and 2 differences
diffFill(x, 1, 2)
}
